user-service/requirements.txt
asyncpg==0.20.1
databases[postgresql]==0.2.6
fastapi==0.48.0
SQLAlchemy==1.3.13
uvicorn==0.11.2
httpx==0.11.1
passlib[bcrypt]==1.7.1
passlib==1.7.1
bcrypt
pyjwt==2.9.0
aioredis==2.0.1